requirements.txt
openai
python-telegram-bot==20.3
requests
flask
python-dotenv